+++ CHANGELOG.md
@@ -4,6 +4,15 @@ All notable changes to this project will be documented in this file.
 
 ## Unreleased
 
+### 2.8.2  - 2021-12-22
+
+---
+
+### Changed 
+
+- Mapper: rename to State
+- Mapper: added exists method
+
 ### 2.8.1  - 2021-12-22
 
 ---